+++ packages/lumx-react/src/components/slideshow/Slideshow.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import range from 'lodash/range';
-import { AspectRatio, Button, FlexBox, ImageBlock, Slideshow, SlideshowItem } from '@lumx/react';
+import { AspectRatio, Button, FlexBox, ImageBlock, Slideshow, SlideshowItem, Orientation } from '@lumx/react';
 import { boolean, number } from '@storybook/addon-knobs';
 import { thumbnailsKnob } from '@lumx/react/stories/knobs/thumbnailsKnob';
 
@@ -15,6 +15,7 @@ export const Simple = ({ theme }: any) => {
 
     return (
         <Slideshow
+            aria-label="Simple carousel example"
             activeIndex={activeIndex}
             autoPlay={autoPlay}
             interval={interval}
@@ -27,7 +28,7 @@ export const Simple = ({ theme }: any) => {
             style={{ width: '50%' }}
         >
             {images.map(({ image, alt }, index) => (
-                <SlideshowItem key={`${image}-${index}`}>
+                <SlideshowItem key={`${image}-${index}`} aria-label={`${index + 1} of ${images.length}`}>
                     <ImageBlock
                         thumbnailProps={{ aspectRatio: AspectRatio.horizontal, loading: 'eager' }}
                         image={image}
@@ -48,6 +49,7 @@ export const SimpleWithAutoPlay = ({ theme }: any) => {
 
     return (
         <Slideshow
+            aria-label="Simple with autoplay example"
             activeIndex={activeIndex}
             autoPlay
             interval={interval}
@@ -61,7 +63,7 @@ export const SimpleWithAutoPlay = ({ theme }: any) => {
             style={{ width: '50%' }}
         >
             {images.map(({ image, alt }, index) => (
-                <SlideshowItem key={`${image}-${index}`}>
+                <SlideshowItem key={`${image}-${index}`} aria-label={`${index + 1} of ${images.length}`}>
                     <ImageBlock
                         thumbnailProps={{ aspectRatio: AspectRatio.horizontal, loading: 'eager' }}
                         image={image}
@@ -75,7 +77,7 @@ export const SimpleWithAutoPlay = ({ theme }: any) => {
 };
 
 export const ResponsiveSlideShowSwipe = () => {
-    const slides = range(3);
+    const slides = range(5);
     return (
         <>
             In responsive mode
@@ -86,14 +88,15 @@ export const ResponsiveSlideShowSwipe = () => {
             </ul>
             <FlexBox vAlign="center">
                 <Slideshow
+                    aria-label="Responsive SlideShow Swipe"
                     activeIndex={0}
                     slideshowControlsProps={{
                         nextButtonProps: { label: 'Next' },
                         previousButtonProps: { label: 'Previous' },
                     }}
                 >
-                    {slides.map((slide) => (
-                        <SlideshowItem key={`${slide}`}>
+                    {slides.map((slide, key) => (
+                        <SlideshowItem key={`${slide}`} aria-label={`${key + 1} of ${slides.length}`}>
                             <FlexBox
                                 style={{ border: '1px solid grey', maxWidth: 300, height: 300 }}
                                 hAlign="center"
@@ -114,3 +117,56 @@ export const ResponsiveSlideShowSwipe = () => {
         </>
     );
 };
+
+const slides = [
+    {
+        src: 'https://www.w3.org/WAI/ARIA/apg/example-index/carousel/images/foyleswarslide__800x600.jpg',
+        alt: 'A man in a suit and fedora and a woman with coiffed hair look sternly into the camera.',
+        title: 'Foyle’s War Revisited',
+        subtitle: '8 pm Sunday, March 8, on TV: Sneak peek at the final season',
+        link: '#',
+    },
+    {
+        src: 'https://www.w3.org/WAI/ARIA/apg/example-index/carousel/images/britcomdavidslide__800x600.jpg',
+        alt: 'British flag with WILL-TV host David Thiel.',
+        title: 'Great Britain Vote: 7 pm Sat.',
+        link: '#',
+    },
+    {
+        src: 'https://www.w3.org/WAI/ARIA/apg/example-index/carousel/images/mag800-2__800x600.jpg',
+        alt: 'Mid-American Gardener panelists on the set.',
+        title: 'Mid-American Gardener: Thursdays at 7 pm',
+        subtitle: 'Watch the latest episode',
+        link: '#',
+    },
+];
+export const WithComplexContent = () => (
+    <Slideshow
+        aria-label="Carousel with complex content"
+        activeIndex={0}
+        groupBy={2}
+        slideshowControlsProps={{
+            nextButtonProps: { label: 'Next' },
+            previousButtonProps: { label: 'Previous' },
+        }}
+        autoPlay
+    >
+        {slides.map((slide, key) => (
+            <SlideshowItem aria-label={`${key + 1} of ${slides.length}`} key={slide.src}>
+                <a href={slide.link}>
+                    <img src={slide.src} alt={slide.alt} />
+                </a>
+                <FlexBox orientation={Orientation.vertical}>
+                    <h3>
+                        <a href={slide.link}>{slide.title}</a>
+                        {/* Add a non focusable element to test that it stays that way after a page change. */}
+                        <button type="button" tabIndex={-1} aria-hidden="true">
+                            Not focusable
+                        </button>
+                    </h3>
+                    {slide.subtitle && <p>{slide.subtitle}</p>}
+                </FlexBox>
+            </SlideshowItem>
+        ))}
+    </Slideshow>
+);